moved code from constants file back to DataContext, adjusted prettier settings
