+++ src/components/Charts/charts.scss
@@ -1,4 +1,4 @@
 #chart {
-	background-color: var(--componentBackground);
-	padding-top: 2rem;
+  background-color: var(--componentBackground);
+  padding-top: 2rem;
 }
