+++ src/components/Header/Header.jsx
@@ -1,7 +1,7 @@
 import './Header.scss';
 
 const Header = () => {
-	return <h1 className='header'>Work / Weather Analizer</h1>;
+  return <h1 className='header'>Work / Weather Analizer</h1>;
 };
 
 export default Header;
