+++ src/components/Header/Header.scss
@@ -1,5 +1,5 @@
 .header {
-	display: flex;
-	justify-content: center;
-	margin-bottom: 30px;
+  display: flex;
+  justify-content: center;
+  margin-bottom: 30px;
 }
