+++ src/components/Inputs/Input.jsx
@@ -1,19 +1,19 @@
 const Input = ({
-	type,
-	value,
-	placeholder,
-	isRequired,
-	handleChange,
-	inputClass,
+  type,
+  value,
+  placeholder,
+  isRequired,
+  handleChange,
+  inputClass,
 }) => (
-	<input
-		type={type}
-		value={value}
-		placeholder={placeholder}
-		required={isRequired}
-		onChange={handleChange}
-		className={inputClass}
-	/>
+  <input
+    type={type}
+    value={value}
+    placeholder={placeholder}
+    required={isRequired}
+    onChange={handleChange}
+    className={inputClass}
+  />
 );
 
 export default Input;
