+++ src/components/Inputs/WeatherButton.jsx
@@ -3,9 +3,9 @@ import { InputContext } from '../../contexts/InputContext';
 import Button from './Button';
 
 const WeatherButton = () => {
-	const { weatherSubmit } = useContext(InputContext);
+  const { weatherSubmit } = useContext(InputContext);
 
-	return <Button onClick={weatherSubmit}>Enter Weather Info</Button>;
+  return <Button onClick={weatherSubmit}>Enter Weather Info</Button>;
 };
 
 export default WeatherButton;
