+++ src/components/Layout/layout.scss
@@ -1,7 +1,7 @@
 .layout {
-	display: flex;
-	flex-direction: column;
-	justify-content: space-between;
-	gap: 2rem;
-	padding: 1rem;
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+  gap: 2rem;
+  padding: 1rem;
 }
