+++ src/components/Warnings/warnings.scss
@@ -1,30 +1,30 @@
 .warning-message-container {
-	display: none;
-	position: absolute;
-	top: 50%;
-	justify-content: center;
-	top: 50%;
-	left: 30%;
+  display: none;
+  position: absolute;
+  top: 50%;
+  justify-content: center;
+  top: 50%;
+  left: 30%;
 
-	& .warning-message {
-		background-color: var(--buttonColor);
-		border-radius: 5px;
-		border: solid #fff 2px;
-		padding: 0 1rem 1rem 1rem;
-		display: grid;
-		grid-template-columns: [first] 3rem [line2] fit-content [line3] 3rem [end];
-	}
+  & .warning-message {
+    background-color: var(--buttonColor);
+    border-radius: 5px;
+    border: solid #fff 2px;
+    padding: 0 1rem 1rem 1rem;
+    display: grid;
+    grid-template-columns: [first] 3rem [line2] fit-content [line3] 3rem [end];
+  }
 
-	& .close-button {
-		grid-column-start: 3;
-		font-size: large;
-		cursor: pointer;
-	}
+  & .close-button {
+    grid-column-start: 3;
+    font-size: large;
+    cursor: pointer;
+  }
 
-	& .warning-text {
-		grid-column-start: 2;
-		display: flex;
-		justify-content: center;
-		cursor: default;
-	}
+  & .warning-text {
+    grid-column-start: 2;
+    display: flex;
+    justify-content: center;
+    cursor: default;
+  }
 }
